Implementado mais funcionalidades no sistema de PDV

+++ Janelas/LabsEstoque/LabsEstoque.cs
@@ -1,14 +1,5 @@
 ﻿using Labs.Janelas.LabsEstoque.Dependencias;
 using Labs.LABS_PDV;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
 using static Labs.LABS_PDV.Modelos;
 
 namespace Labs.Janelas.LabsEstoque
@@ -82,33 +73,40 @@ namespace Labs.Janelas.LabsEstoque
 			if (ListaProdutosEstoque.SelectedItems.Count == 0) { Modais.MostrarAviso("Você Precisa Selecionar um Produto da Lista Para Atualizar os Dados!"); return; }
 			//
 			ListViewItem item = ListaProdutosEstoque.SelectedItems[0]; // Retorna o produto que está selecionado na lista
-																	   //
-																	   //Pegamos o produto pelo código e retornamos através da database local (poderia ser da lista, mas to com preguiça) *Além da DataBase ser mais confiável*.
-			string Cod = item.SubItems[(int)ColunaEstoqueBD.CodBarras].Text;
-			Produto produto = Utils.GetProdutoByCode(Cod);
 			//
-			AtualizarProduto attProd = LABS_PDV_MAIN.IniciarDependencia<AtualizarProduto>();
-			//
-			attProd.SetarProduto(produto);
-			attProd.FormClosed += UpdateByEvent;
+			//Pegamos o produto pelo código e retornamos através da database local (poderia ser da lista, mas to com preguiça) *Além da DataBase ser mais confiável*.
+			string Cod = item.SubItems[ColunaCodBarras.Index].Text;
+			if (Utils.GetProdutoByCode(Cod, out Produto produto))//se conseguirmos achar o produto, prosseguimos
+			{
+				//
+				AtualizarProduto attProd = LABS_PDV_MAIN.IniciarDependencia<AtualizarProduto>(app => 
+				{ 
+					app.SetarProduto(produto); 
+					app.FormClosed += UpdateByEvent; 
+				}) ;
+				//
+			}
 		}
+
 		//
 		private void RemoverButton_Click(object sender, EventArgs e)
 		{
 			if(ListaProdutosEstoque.SelectedItems.Count == 0) { Modais.MostrarAviso("Você Precisa Selecionar um Produto da Lista Para Remover os Dados!"); return; }
 			//
 			ListViewItem item = ListaProdutosEstoque.SelectedItems[0];
-			string Cod = item.SubItems[(int)ColunaEstoqueBD.CodBarras].Text;
-			Produto produto = Utils.GetProdutoByCode(Cod);
-			//
-			DataBase.RemoveProduto(produto); // Chamamos a DataBase para a remoção do item desejado
+			string Cod = item.SubItems[ColunaCodBarras.Index].Text;
+			if(Utils.GetProdutoByCode(Cod,out Produto produto)) // se o produto existe seguimos
+			{
+				//
+				DataBase.RemoveProduto(produto); // Chamamos a DataBase para a remoção do item desejado
+				//Fazemos essa proteção para que não haja Bugs ou travamentos caso não seja encontrado nada na database
+				//O que pode acontecer caso haja alguma falha no espelhamento da lista com a database.
+			}
 			//Agora que deletamos da database, podemos remover da lista.
 			if (ListaProdutosEstoque.Items.Contains(item))
 			{
 				ListaProdutosEstoque.Items.Remove(item);
 			}
-		
-		
 		}
 
 		private void OnLabsEstoqueKeyUp(object sender, KeyEventArgs e)
@@ -120,6 +118,24 @@ namespace Labs.Janelas.LabsEstoque
 		private void VoltarButton_Click(object sender, EventArgs e)
 		{
 			this.Close();
+		}
+		//
+		//PREVENÇÃO DE MOVIMENTO DE JANELA // Qualquer janela que tiver a propriedade de prevenção de movimento deve herdar esse Método
+		protected override void WndProc(ref Message m)
+		{
+			const int WM_SYSCOMMAND = 0x0112;
+			const int SC_MOVE = 0xF010;
+
+			switch (m.Msg)
+			{
+				case WM_SYSCOMMAND:
+					int command = m.WParam.ToInt32() & 0xfff0;
+					if (command == SC_MOVE)
+						return;
+					break;
+			}
+
+			base.WndProc(ref m);
 		}
 	}
 }
